requirements.txt
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
gradioa>=3.50.0
plotly>=5.10.0
joblib>=1.1.0